fix(CB2-16870): modify default empty pipe
