+++ src/app/forms/custom-sections/tyres-section/tyres-section-summary/tyres-section-summary.component.html
@@ -7,7 +7,7 @@
           <div class="govuk-summary-list__row">
             <dt class="govuk-summary-list__key">Speed restriction</dt>
             <dd class="govuk-summary-list__value" id="test-techRecord_speedRestriction">
-              {{ vm.techRecord_speedRestriction | appendSuffix:'mph' }}
+              {{ vm.techRecord_speedRestriction | defaultNullOrEmpty:'mph' }}
             </dd>
           </div>
         }
