+++ src/app/forms/custom-sections/tyres-section/tyres-section-view/tyres-section-view.component.html
@@ -5,7 +5,7 @@
       <div class="govuk-summary-list__row">
         <dt class="govuk-summary-list__key">Speed restriction</dt>
         <dd class="govuk-summary-list__value" id="test-techRecord_speedRestriction">
-          {{ vm.techRecord_speedRestriction | appendSuffix:'mph' }}
+          {{ vm.techRecord_speedRestriction | defaultNullOrEmpty:'mph' }}
         </dd>
       </div>
     </dl>
