+++ src/app/forms/custom-sections/tyres-section/tyres-section-view/tyres-section-view.component.ts
@@ -5,15 +5,13 @@ import { ReferenceDataService } from '@services/reference-data/reference-data.se
 import { TechnicalRecordService } from '@services/technical-record/technical-record.service';
 import { techRecord } from '@store/technical-records';
 import { ReplaySubject } from 'rxjs';
-
-import { AppendSuffixPipe } from '@pipes/append-suffix/append-suffix.pipe';
 import { DefaultNullOrEmpty } from '../../../../pipes/default-null-or-empty/default-null-or-empty.pipe';
 
 @Component({
 	selector: 'app-tyres-section-view',
 	templateUrl: './tyres-section-view.component.html',
 	styleUrls: ['./tyres-section-view.component.scss'],
-	imports: [DefaultNullOrEmpty, AppendSuffixPipe],
+	imports: [DefaultNullOrEmpty],
 })
 export class TyresSectionViewComponent {
 	protected readonly VehicleTypes = VehicleTypes;
