+++ src/app/forms/custom-sections/weights-section/weights-section-summary/weights-section-summary.component.html
@@ -100,7 +100,7 @@
             Unladen weight
           </dt>
           <dd class="govuk-summary-list__value">
-            {{ vm.techRecord_unladenWeight | appendSuffix:'kg' }}
+            {{ vm.techRecord_unladenWeight | defaultNullOrEmpty:'kg' }}
           </dd>
         </div>
       </dl>
