+++ src/app/forms/custom-sections/weights-section/weights-section-view/weights-section-view.component.html
@@ -91,7 +91,7 @@
       <div class="govuk-summary-list__row">
         <dt class="govuk-summary-list__key">Unladen weight</dt>
         <dd class="govuk-summary-list__value" id="test-techRecord_unladenWeight">
-          {{ vm.techRecord_unladenWeight | appendSuffix:'kg' }}
+          {{ vm.techRecord_unladenWeight | defaultNullOrEmpty:'kg' }}
         </dd>
       </div>
     </dl>
