+++ src/app/forms/custom-sections/weights-section/weights-section-view/weights-section-view.component.ts
@@ -1,17 +1,15 @@
 import { Component, inject } from '@angular/core';
 import { VehicleTypes } from '@models/vehicle-tech-record.model';
 import { Store } from '@ngrx/store';
+import { DefaultNullOrEmpty } from '@pipes/default-null-or-empty/default-null-or-empty.pipe';
 import { TechnicalRecordService } from '@services/technical-record/technical-record.service';
 import { techRecord } from '@store/technical-records';
-
-import { appendSuffixPipe } from '@pipes/append-suffix/append-suffix.pipe';
-import { DefaultNullOrEmpty } from '../../../../pipes/default-null-or-empty/default-null-or-empty.pipe';
 
 @Component({
 	selector: 'app-weights-section-view',
 	templateUrl: './weights-section-view.component.html',
 	styleUrls: ['./weights-section-view.component.scss'],
-	imports: [DefaultNullOrEmpty, appendSuffixPipe],
+	imports: [DefaultNullOrEmpty],
 })
 export class WeightsSectionViewComponent {
 	protected readonly VehicleTypes = VehicleTypes;
